archsigのテスト配置をモジュール境界へ整理
