+++ tools/archsig/src/dataset.rs
@@ -341,3 +341,100 @@ fn metric_status_reason(axis: &str, metric_status: &BTreeMap<String, MetricStatu
         .and_then(|status| status.reason.clone())
         .unwrap_or_else(|| "metricStatus entry is missing".to_string())
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::EMPIRICAL_DATASET_SCHEMA_VERSION;
+    use crate::test_support::{dataset_input, sig0_document_for_edges};
+
+    use super::build_empirical_dataset;
+
+    #[test]
+    fn builds_empirical_dataset_with_comparable_and_unmeasured_deltas() {
+        let before = sig0_document_for_edges(
+            vec![
+                ("A", "A.lean"),
+                ("B", "B.lean"),
+                ("C", "C.lean"),
+                ("D", "D.lean"),
+            ],
+            vec![("A", "B")],
+        );
+        let after = sig0_document_for_edges(
+            vec![
+                ("A", "A.lean"),
+                ("B", "B.lean"),
+                ("C", "C.lean"),
+                ("D", "D.lean"),
+            ],
+            vec![("A", "B"), ("A", "C"), ("C", "D")],
+        );
+
+        let dataset = build_empirical_dataset(&before, &after, dataset_input(), "head")
+            .expect("dataset builds");
+
+        assert_eq!(dataset.schema_version, EMPIRICAL_DATASET_SCHEMA_VERSION);
+        assert_eq!(dataset.signature_before.commit.sha, "base-sha");
+        assert_eq!(dataset.signature_after.commit.role, "head");
+        assert_eq!(dataset.delta_signature_signed.fanout_risk, Some(2));
+        assert_eq!(dataset.delta_signature_signed.max_fanout, Some(1));
+        assert_eq!(dataset.delta_signature_signed.reachable_cone_size, Some(2));
+        assert_eq!(
+            dataset.delta_signature_signed.boundary_violation_count,
+            None
+        );
+        assert_eq!(dataset.signature_after.signature.weighted_scc_risk, None);
+        assert!(
+            !dataset
+                .signature_after
+                .metric_status
+                .get("weightedSccRisk")
+                .expect("weighted status")
+                .measured
+        );
+        assert_eq!(dataset.signature_after.signature.runtime_propagation, None);
+        assert!(
+            !dataset
+                .signature_after
+                .metric_status
+                .get("runtimePropagation")
+                .expect("runtime status")
+                .measured
+        );
+
+        let fanout_status = dataset
+            .metric_delta_status
+            .get("fanoutRisk")
+            .expect("fanout delta status");
+        assert!(fanout_status.comparable);
+        assert!(fanout_status.before_measured);
+        assert!(fanout_status.after_measured);
+
+        let boundary_status = dataset
+            .metric_delta_status
+            .get("boundaryViolationCount")
+            .expect("boundary delta status");
+        assert!(!boundary_status.comparable);
+        assert!(!boundary_status.before_measured);
+        assert!(!boundary_status.after_measured);
+        assert!(
+            boundary_status
+                .reason
+                .contains("policy file not provided before")
+                || boundary_status.reason.contains("policy file not provided")
+        );
+    }
+
+    #[test]
+    fn dataset_merge_role_uses_merge_commit() {
+        let document = sig0_document_for_edges(vec![("A", "A.lean")], Vec::new());
+        let dataset = build_empirical_dataset(&document, &document, dataset_input(), "merge")
+            .expect("dataset builds");
+
+        assert_eq!(dataset.signature_after.commit.sha, "merge-sha");
+        assert_eq!(
+            dataset.analysis_metadata.signature_after_commit_role,
+            "merge"
+        );
+    }
+}
